fix deletelist reflex

+++ app/reflexes/todo_list_reflex.rb
@@ -2,7 +2,8 @@ class TodoListReflex < ApplicationReflex
 
     def delete_that_list
       todo_list = TodoList.find(element.dataset.id)
-      todo_list.destroy
+      items = TodoItem.where(todo_list_id: todo_list)
+      items.destroy_all && todo_list.destroy
     end
   
 end
